Tags, line numbers, and assorted refactoring
This commit introduces a new concept called a TAG, which is used to
annotate lines.  TAGs are computed as needed, and store any
line-specific information that depends on previous lines.

Syntax info as for exp-syntax.lisp has been moved into tags.

In addition, the tag stores "real" lines numbers for display to the user
(as opposed to LINE-NUMBER, which can have gaps).  At the moment, only
the Qt backend actually displays these new line numbers.

The third piece of information in the tag stores the character offset of
the line from the beginning of the file.  We support this "because we
can", not because it's actually needed yet.

The update of TAGs isn't entirely correct yet, but is on par with the
previous syntax info implementation, which has always missed certain
updates, because only lines that actually changed also got new
information.  The whole syntax and tag concept, of course, must be
centered around the notion that a line update can affect other lines
further down, so that is not sufficient.

The current workaround is that the Qt backend asks for tag computation
before showing the line number; obviously this needs to be moved to a
better place.  It also implies that only the Qt backend has correct
syntax highlighting updates as the moment.  In return though, its
performance suffers, so perhaps it is good that other backends don't
have the same "feature" yet.

Also needed in the future will be a limit, which should allow users to
stop tag computation based on a number of lines or character, just like
emacs font locking stops at some point.  (The limit would be easy to
implement at computation time, but all places asking for tags will need
to be updated to deal with the NIL tags that would then occur.)