Update to use command-line-arguments v2.0.0
The command-line-arguments package version 2 changed it's handling of
arguments in a backwards incompatible way.  The return value of any
:action functions defined on arguments are now saved as the new value
of the argument in the returned `*command-line-arguments*'.

This commit updates hemlock so that it defines and ignores these new
keyword arguments when they are passed through to the `hemlock' and
`%start-slave' functions.  This preservers the old behavior.